requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
celery==5.3.6
redis==5.0.3
sqlmodel==0.0.16
python-dotenv==1.0.1
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
prophet==1.1.5
python-multipart==0.0.6
beautifulsoup4==4.12.3
httpx==0.27.0
tenacity==8.2.3
structlog==24.1.0
playwright==1.45.0
